OB-78 AC-H2: widen transaction nonce int16 -> int64 (CONSENSUS / GENESIS RESET)
The nonce is part of the signed TxParam. int16 allowed only 32768 values,
weakening replay protection. Widen to int64: serialize as 8 bytes, shift
MultiSignTx to offset 39, bump the min TxParam length to 40, and update
genesis and all TxParam construction sites.

BREAKING: changes the signed transaction layout, so every existing
signature/hash (including genesis) is invalidated. This branch starts a
fresh chain from a new genesis. Weak math/rand nonce generation at these
sites is addressed separately (WH-M2).

Adds a TxParam round-trip test covering nonces beyond the old int16 range.

+++ cmd/gui/qtwidgets/dex.go
@@ -555,7 +555,7 @@ func MakeTransaction(sender, coinAddr common.Address, primary bool) {
 			ChainID:     ChainID,
 			Sender:      sender,
 			SendingTime: common.GetCurrentTimeStampInSecond(),
-			Nonce:       int16(rand.Intn(0xffff)),
+			Nonce:       int64(rand.Intn(0xffff)),
 		}
 		tx := transactionsDefinition.Transaction{
 			TxData:          txd,
@@ -660,7 +660,7 @@ func MakeTrade(sender, coinAddr common.Address, isBuy bool, primary bool) {
 			ChainID:     ChainID,
 			Sender:      sender,
 			SendingTime: common.GetCurrentTimeStampInSecond(),
-			Nonce:       int16(rand.Intn(0xffff)),
+			Nonce:       int64(rand.Intn(0xffff)),
 		}
 		tx := transactionsDefinition.Transaction{
 			TxData:          txd,
